requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
beautifulsoup4
requests
numpy
pydantic
